+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/settings/censorship/components/StealthApiSection.kt
@@ -57,6 +57,7 @@ fun StealthApiSection(shape: Shape = RoundedCornerShape(8.dp)) {
 				Text(
 					text = stringResource(R.string.censorship_stealth_title),
 					style = MaterialTheme.typography.titleMedium,
+					fontFamily = FontFamily(Font(R.font.lab_grotesque_regular)),
 				)
 				ScaledSwitch(
 					checked = true,