+++ nym-vpn-apple/Settings/Sources/Settings/AppSettingsSectionKind.swift
@@ -8,5 +8,6 @@ public enum AppSettingsSectionKind: SettingsSectionKind {
     case feedback
     case killSwitch
     case legal
+    case systemStatus
     case logout
 }