+++ nym-vpn-android/CHANGELOG.md
@@ -9,12 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Node details added in quick settings notification (https://github.com/nymtech/nym-vpn-client/pull/4029)
-- Android: Bi-mode Split Tunneling
-
+- AmneziaWG section added to censorship screen (https://github.com/nymtech/nym-vpn-client/pull/4033)
+- Revoke implementation added for VPN service (https://github.com/nymtech/nym-vpn-client/pull/4041)
+- Android: Bi-mode Split Tunneling (https://github.com/nymtech/nym-vpn-client/pull/4049
+- 
 ### Changed
 
 ### Fixed
 - Connection error notification cleared after successfully connecting (https://github.com/nymtech/nym-vpn-client/pull/4029)
+- Fix connection drops after language change (https://github.com/nymtech/nym-vpn-client/pull/4034)
+- Fix QUIC autostart issue (https://github.com/nymtech/nym-vpn-client/pull/4040)
 
 ## [2.4.0]